Moved each effect to a separate file
