+++ Client/Effect/IEffect.cs
@@ -1,0 +1,12 @@
+﻿using System.Threading.Tasks;
+
+namespace PocceMod.Client.Effect
+{
+    public interface IEffect
+    {
+        string Key { get; }
+        bool Expired { get; }
+        Task Init();
+        void Clear();
+    }
+}
